+++ city_infromation_service/src/main/java/viachaslausviatski/city_infromation_service/repositories/PopularityRepository.java
@@ -3,5 +3,5 @@ package viachaslausviatski.city_infromation_service.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import viachaslausviatski.city_infromation_service.entities.Popularity;
 
-public interface PopularityRepository extends JpaRepository<Popularity,Integer> {
+public interface PopularityRepository extends JpaRepository<Popularity,Long> {
 }
